[MOVE]: MOVE #416 OAuthToken,AuthResponse into core/network
