--- app/src/main/java/org/sopt/official/data/mapper/AuthResponseMapper.kt
@@ -1,13 +1,0 @@
-package org.sopt.official.data.mapper
-
-import org.sopt.official.network.model.response.AuthResponse
-
-fun toAuthResponse(responseItem: AuthResponse): org.sopt.official.data.model.response.AuthResponse {
-    return org.sopt.official.data.model.response.AuthResponse(
-        accessToken = responseItem.accessToken,
-        refreshToken = responseItem.refreshToken,
-        playgroundToken = responseItem.playgroundToken,
-        status = responseItem.status
-    )
-}
-
