+++ app/src/main/java/org/sopt/official/data/source/api/auth/RemoteAuthDataSource.kt
@@ -25,7 +25,7 @@
 package org.sopt.official.data.source.api.auth
 
 import org.sopt.official.data.model.request.LogOutRequest
-import org.sopt.official.data.model.response.AuthResponse
+import org.sopt.official.network.model.response.AuthResponse
 import org.sopt.official.data.model.response.LogOutResponse
 import org.sopt.official.network.model.request.RefreshRequest
