+++ app/src/main/java/org/sopt/official/util/Auth.kt
@@ -24,7 +24,7 @@
  */
 package org.sopt.official.util
 
-import org.sopt.official.auth.data.remote.model.response.OAuthToken
+import org.sopt.official.network.model.response.OAuthToken
 import org.sopt.official.domain.entity.auth.Auth
 import org.sopt.official.domain.entity.auth.Token
 import org.sopt.official.domain.entity.auth.UserStatus
