+++ core/network/src/main/java/org/sopt/official/network/service/RefreshService.kt
@@ -1,7 +1,8 @@
 package org.sopt.official.network.service
 
-import org.sopt.official.network.model.response.AuthResponse
+
 import org.sopt.official.network.model.request.RefreshRequest
+import org.sopt.official.network.model.response.AuthResponse
 import retrofit2.http.Body
 import retrofit2.http.PATCH
