+++ feature/auth/build.gradle.kts
@@ -34,6 +34,7 @@ android {
 
 dependencies {
     implementation(projects.core.common)
+    implementation(projects.core.network)
     implementation(libs.customtab)
     implementation(libs.retrofit)
     implementation(libs.retrofit.kotlin.serialization.converter)
