+++ feature/auth/src/main/java/org/sopt/official/auth/PlaygroundAuth.kt
@@ -33,7 +33,7 @@ import kotlinx.coroutines.launch
 import org.sopt.official.auth.data.PlaygroundApiManager
 import org.sopt.official.auth.data.PlaygroundAuthDatasource
 import org.sopt.official.auth.data.RemotePlaygroundAuthDatasource
-import org.sopt.official.auth.data.remote.model.response.OAuthToken
+import org.sopt.official.network.model.response.OAuthToken
 import org.sopt.official.auth.utils.PlaygroundUriProvider
 
 object PlaygroundAuth {
