+++ feature/auth/src/main/java/org/sopt/official/auth/data/PlaygroundAuthDatasource.kt
@@ -24,7 +24,7 @@
  */
 package org.sopt.official.auth.data
 
-import org.sopt.official.auth.data.remote.model.response.OAuthToken
+import org.sopt.official.network.model.response.OAuthToken
 
 interface PlaygroundAuthDatasource {
     suspend fun oauth(code: String): Result<OAuthToken>
