+++ feature/auth/src/main/java/org/sopt/official/auth/data/remote/AuthService.kt
@@ -25,7 +25,7 @@
 package org.sopt.official.auth.data.remote
 
 import org.sopt.official.auth.data.remote.model.request.RequestToken
-import org.sopt.official.auth.data.remote.model.response.OAuthToken
+import org.sopt.official.network.model.response.OAuthToken
 import retrofit2.http.Body
 import retrofit2.http.POST
